+++ database/seeds/seed_demo_peripheral_catalogue.php
@@ -148,7 +148,7 @@ try {
             if (in_array($definition['data_type'], ['integer', 'decimal'], true)) {
                 $row['value_number'] = $value;
             } elseif ($definition['data_type'] === 'boolean') {
-                $row['value_boolean'] = (bool) $value;
+                $row['value_boolean'] = (int) (bool) $value;
             } elseif ($definition['data_type'] === 'option') {
                 $row['option_id'] = $options[$code][(string) $value] ?? null;
                 if ($row['option_id'] === null) {
